+++ src/int/audiof.h
@@ -2,6 +2,7 @@
 #define FALLOUT_INT_AUDIOF_H_
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #include "sound_decoder.h"
 
@@ -12,7 +13,7 @@ typedef enum AudioFileFlags {
 
 typedef struct AudioFile {
     int flags;
-    int fileHandle;
+    intptr_t fileHandle;
     SoundDecoder* soundDecoder;
     int fileSize;
     int field_10;
@@ -22,13 +23,13 @@ typedef struct AudioFile {
 
 typedef bool(AudioFileIsCompressedProc)(char* filePath);
 
-int audiofOpen(const char* fname, int flags, ...);
-int audiofCloseFile(int a1);
-int audiofRead(int a1, void* buf, unsigned int size);
-long audiofSeek(int handle, long offset, int origin);
-long audiofFileSize(int a1);
-long audiofTell(int a1);
-int audiofWrite(int handle, const void* buf, unsigned int size);
+intptr_t audiofOpen(const char* fname, int flags, ...);
+int audiofCloseFile(intptr_t a1);
+int audiofRead(intptr_t a1, void* buf, unsigned int size);
+long audiofSeek(intptr_t handle, long offset, int origin);
+long audiofFileSize(intptr_t a1);
+long audiofTell(intptr_t a1);
+int audiofWrite(intptr_t handle, const void* buf, unsigned int size);
 int initAudiof(AudioFileIsCompressedProc* isCompressedProc);
 void audiofClose();
